feat: support watching multiple stack operations in parallel
Multiple stacks can be watched by piping the output of more than
one aws cloudformation command into the stdin of cfn-monitor. To
support this use case, the stack name will be prepended to the
status lines and colors will be used to differentiate events for
different stacks (if terminal supports those).