fix: completion detection when cleanup occurs
Previously we could get the following events in this order:

* UPDATE_COMPLETE
* UPDATE_CLEANUP_IN_PROGRESS

When processed, the complete flag was set based on the latter
event (earlier time wise), not the first event (later time wise).

This commit fixes completion detection to consider stack complete
if any event indicates it is complete.

This fixes an issue with cfn-monitor hanging in certain edge cases.